chore: Make toggleWidget a global function
